Fixing more warnings from -Wall.

+++ src/rendering/eavl2DAxisAnnotation.h
@@ -97,7 +97,7 @@ class eavl2DAxisAnnotation : public eavlAnnotation
     void SetLabelFontScale(float s)
     {
         fontscale = s;
-        for (int i=0; i<labels.size(); i++)
+        for (unsigned int i=0; i<labels.size(); i++)
             labels[i]->SetScale(s);
     }
     void SetRangeForAutoTicks(double l, double u)
@@ -147,7 +147,7 @@ class eavl2DAxisAnnotation : public eavlAnnotation
 
         // major ticks
         int nmajor = maj_proportions.size();
-        while (labels.size() < nmajor)
+        while ((int)labels.size() < nmajor)
         {
             if (worldSpace)
             {
